+++ src/Tests/EventRouting/SubscriptionEntryJsonConverterTests.cs
@@ -4,113 +4,6 @@ using System.Text.Json;
 using NUnit.Framework;
 
 [TestFixture]
-public class SubscriptionEntryTests
-{
-    [Test]
-    public void Implicit_conversion_from_string_creates_default_entry()
-    {
-        SubscriptionEntry entry = "MyTopic";
-
-        using (Assert.EnterMultipleScope())
-        {
-            Assert.That(entry.Topic, Is.EqualTo("MyTopic"));
-            Assert.That(entry.RoutingMode, Is.Null);
-        }
-    }
-
-    [Test]
-    public void Can_create_with_explicit_routing_mode()
-    {
-        var entry = new SubscriptionEntry("MyTopic", TopicRoutingMode.CorrelationFilter);
-
-        using (Assert.EnterMultipleScope())
-        {
-            Assert.That(entry.Topic, Is.EqualTo("MyTopic"));
-            Assert.That(entry.RoutingMode, Is.EqualTo(TopicRoutingMode.CorrelationFilter));
-        }
-    }
-
-    [Test]
-    public void Default_routing_mode_is_null()
-    {
-        var entry = new SubscriptionEntry("MyTopic");
-
-        Assert.That(entry.RoutingMode, Is.Null);
-    }
-
-    [Test]
-    public void Can_use_all_routing_modes()
-    {
-        var notMultiplexedEntry = new SubscriptionEntry("Topic1", TopicRoutingMode.NotMultiplexed);
-        var correlationEntry = new SubscriptionEntry("Topic2", TopicRoutingMode.CorrelationFilter);
-        var sqlEntry = new SubscriptionEntry("Topic3", TopicRoutingMode.SqlLikeFilter);
-
-        using (Assert.EnterMultipleScope())
-        {
-            Assert.That(notMultiplexedEntry.RoutingMode, Is.EqualTo(TopicRoutingMode.NotMultiplexed));
-            Assert.That(correlationEntry.RoutingMode, Is.EqualTo(TopicRoutingMode.CorrelationFilter));
-            Assert.That(sqlEntry.RoutingMode, Is.EqualTo(TopicRoutingMode.SqlLikeFilter));
-        }
-    }
-}
-
-[TestFixture]
-public class RoutingOptionsTests
-{
-    [Test]
-    public void Default_values_are_correct()
-    {
-        var options = new RoutingOptions();
-
-        Assert.That(options.Mode, Is.Null);
-    }
-
-    [Test]
-    public void Can_set_correlation_routing_mode()
-    {
-        var options = new RoutingOptions { Mode = TopicRoutingMode.CorrelationFilter };
-
-        Assert.That(options.Mode, Is.EqualTo(TopicRoutingMode.CorrelationFilter));
-    }
-
-    [Test]
-    public void Can_set_sql_routing_mode()
-    {
-        var options = new RoutingOptions { Mode = TopicRoutingMode.SqlLikeFilter };
-
-        Assert.That(options.Mode, Is.EqualTo(TopicRoutingMode.SqlLikeFilter));
-    }
-}
-
-[TestFixture]
-public class FallbackTopicOptionsTests
-{
-    [Test]
-    public void Default_values_are_correct()
-    {
-        var options = new FallbackTopicOptions();
-
-        using (Assert.EnterMultipleScope())
-        {
-            Assert.That(options.TopicName, Is.Null);
-            Assert.That(options.RoutingMode, Is.Null);
-        }
-    }
-
-    [Test]
-    public void Can_set_topic_name_and_mode()
-    {
-        var options = new FallbackTopicOptions { TopicName = "SharedTopic", RoutingMode = TopicRoutingMode.CorrelationFilter };
-
-        using (Assert.EnterMultipleScope())
-        {
-            Assert.That(options.TopicName, Is.EqualTo("SharedTopic"));
-            Assert.That(options.RoutingMode, Is.EqualTo(TopicRoutingMode.CorrelationFilter));
-        }
-    }
-}
-
-[TestFixture]
 public class SubscriptionEntryJsonConverterTests
 {
     [Test]
